serverless server

+++ server.js
@@ -1,34 +1,31 @@
-const express       = require('express');
-const cors          = require('cors');
-const dotenv        = require('dotenv');
-const passport      = require('passport');
-const passportJWT   = require('passport-jwt');
-const jwt           = require('jsonwebtoken');
-const userService   = require('./user-service.js');
+// server.js
+const serverless   = require('serverless-http');
+const express      = require('express');
+const cors         = require('cors');
+const dotenv       = require('dotenv');
+const passport     = require('passport');
+const passportJWT  = require('passport-jwt');
+const jwt          = require('jsonwebtoken');
+const userService  = require('./user-service.js');
 
 dotenv.config();
-const HTTP_PORT = process.env.PORT || 8080;
 
 const ExtractJwt  = passportJWT.ExtractJwt;
 const JwtStrategy = passportJWT.Strategy;
-
-const jwtOptions = {
+const jwtOptions  = {
   jwtFromRequest: ExtractJwt.fromAuthHeaderWithScheme('jwt'),
   secretOrKey:    process.env.JWT_SECRET
 };
 
-const strategy = new JwtStrategy(jwtOptions, (jwt_payload, next) => {
+passport.use(new JwtStrategy(jwtOptions, (jwt_payload, next) => {
   if (jwt_payload && jwt_payload._id && jwt_payload.userName) {
     next(null, { _id: jwt_payload._id, userName: jwt_payload.userName });
   } else {
     next(null, false);
   }
-});
-
-passport.use(strategy);
+}));
 
 const app = express();
-
 app.use(passport.initialize());
 app.use(express.json());
 app.use(cors());
@@ -43,7 +40,7 @@ app.post("/api/user/login", (req, res) => {
   userService.checkUser(req.body)
     .then(user => {
       const payload = { _id: user._id, userName: user.userName };
-      const token = jwt.sign(payload, process.env.JWT_SECRET, { expiresIn: '2h' });
+      const token   = jwt.sign(payload, process.env.JWT_SECRET, { expiresIn: '2h' });
       res.json({ message: 'login successful', token });
     })
     .catch(err => res.status(422).json({ message: err }));
@@ -51,49 +48,21 @@ app.post("/api/user/login", (req, res) => {
 
 const auth = passport.authenticate('jwt', { session: false });
 
-app.get("/api/user/favourites", auth, (req, res) => {
-  userService.getFavourites(req.user._id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
+app.get   ("/api/user/favourites",        auth, (req, res) => userService.getFavourites (req.user._id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
+app.put   ("/api/user/favourites/:id",    auth, (req, res) => userService.addFavourite   (req.user._id, req.params.id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
+app.delete("/api/user/favourites/:id",    auth, (req, res) => userService.removeFavourite(req.user._id, req.params.id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
+app.get   ("/api/user/history",           auth, (req, res) => userService.getHistory   (req.user._id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
+app.put   ("/api/user/history/:id",       auth, (req, res) => userService.addHistory     (req.user._id, req.params.id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
+app.delete("/api/user/history/:id",       auth, (req, res) => userService.removeHistory  (req.user._id, req.params.id).then(data => res.json(data)).catch(err => res.status(422).json({ error: err })));
 
-app.put("/api/user/favourites/:id", auth, (req, res) => {
-  userService.addFavourite(req.user._id, req.params.id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
+//
+// Connect to Mongo once at cold start, then export the handler
+//
+let connectPromise = userService.connect();
 
-app.delete("/api/user/favourites/:id", auth, (req, res) => {
-  userService.removeFavourite(req.user._id, req.params.id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
+const handler = serverless(app);
 
-app.get("/api/user/history", auth, (req, res) => {
-  userService.getHistory(req.user._id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
-
-app.put("/api/user/history/:id", auth, (req, res) => {
-  userService.addHistory(req.user._id, req.params.id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
-
-app.delete("/api/user/history/:id", auth, (req, res) => {
-  userService.removeHistory(req.user._id, req.params.id)
-    .then(data => res.json(data))
-    .catch(err => res.status(422).json({ error: err }));
-});
-
-userService.connect()
-  .then(() => {
-    app.listen(HTTP_PORT, () => {
-      console.log(`API listening on: ${HTTP_PORT}`);
-    });
-  })
-  .catch(err => {
-    console.error("unable to start the server:", err);
-    process.exit(1);
-  });
+module.exports = async (req, res) => {
+  await connectPromise;
+  return handler(req, res);
+};
